backend/requirements.txt
Django==5.1.0
djangorestframework==3.15.2
djoser==2.2.3
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.4.0
python-decouple==3.8
Pillow==10.4.0
psycopg2-binary==2.9.9